pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "qenex-pathfinder"
version = "0.1.0"
description = "Security audit tool for Python codebases — CLI + MCP server"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = ["pyyaml>=6.0"]

[project.optional-dependencies]
mcp = ["mcp[cli]>=1.0"]
dev = ["pytest>=7.0", "pytest-cov"]

[project.scripts]
pathfinder = "pathfinder.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
